feat(node): Send client reports
